export query-chunks
