+++ src/index.ts
@@ -1,3 +1,6 @@
 export type * from './promptql';
 export * from './types';
 export * from './client';
+export * from './query-chunks';
+
+export { withActiveSpan } from './utils';
